Fix real evict example
