+++ synthesizer/process/src/lib.rs
@@ -217,7 +217,6 @@ impl<N: Network> Process<N> {
         Ok(())
     }
 
-    #[cfg(test)]
     /// Returns the size of the cache.
     #[inline]
     pub fn num_stacks_in_memory(&self) -> usize {
